+++ lib/GUI/Bar.hpp
@@ -5,12 +5,12 @@
 
 class Bar : GUI {
     uint8_t height, width;
-    uint8_t maxCapacity, currentCapacity;
+    uint16_t maxCapacity, currentCapacity;
 
     void draw() override;
 
     public:
-        Bar(Adafruit_SSD1306 *displayDriver, uint8_t posX, uint8_t posY, uint8_t width, uint8_t height, uint8_t currentCapacity, uint8_t maxCapacity);
+        Bar(Adafruit_SSD1306 *displayDriver, uint8_t posX, uint8_t posY, uint8_t width, uint8_t height, uint16_t currentCapacity, uint16_t maxCapacity);
 
         void setCapacity(uint8_t newCapacity);
 };
